# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Orion Belt API
  description: |
    REST and WebSocket API for the Orion Belt privileged access gateway.

    **Base URL:** `/api/v1` (plus `/health` and optional `/metrics` outside the versioned prefix).

    **Authentication** (protected routes — try in order):
    1. `X-API-Key` header
    2. `X-Session-Token` header, `session_token` cookie, or `token` / `access_token` query (WebSocket)
    3. `Authorization: Bearer <JWT>`

    Pubkey login (`/public/login*`) requires a prior `POST /public/auth/challenge` and a signature proving possession of the private key.
    Password login: `POST /public/login/password` (password + TOTP). Console also supports WebAuthn and `osh login` browser bootstrap.

    Optional **SSH CA** endpoints issue short-lived User certs and expose Host CA trust material (`docs/SSH_CA.md`).

    **Admin routes** (`/api/v1/admin/*`) additionally require role `admin` or `operator` (or `is_admin`).

    **Rate limit:** protected (non-admin) routes — default 600 req/min per user (`auth.rate_limit_per_minute`).
  version: "0.8.1"
  contact:
    name: Orion Belt Maintainers
    url: https://github.com/zrougamed/orion-belt
  license:
    name: Apache-2.0 WITH Commons-Clause
    url: https://github.com/zrougamed/orion-belt/blob/master/LICENSE
servers:
  - url: http://localhost:8080
    description: Local / lab default
  - url: "{scheme}://{host}"
    description: Deployed gateway
    variables:
      scheme:
        default: https
...
paths:
...
  /api/v1/webauthn/register/finish:
    post:
      tags: [WebAuthn]
      summary: Finish WebAuthn registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  id: { type: string }
...
  /api/v1/webauthn/credentials/{id}:
    delete:
      tags: [WebAuthn]
      summary: Delete WebAuthn credential
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    SessionHeader:
      type: apiKey
      in: header
      name: X-Session-Token
    SessionCookie:
      type: apiKey
      in: cookie
      name: session_token
    SessionQuery:
      type: apiKey
      in: query
      name: token
    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    Message:
      type: object
      properties:
        message: { type: string }
...
security:
  - ApiKeyAuth: []
  - SessionHeader: []
  - SessionCookie: []
  - BearerJWT: []